Finish writing passport validation method
